requirements.txt
anaconda-anon-usage @ file:///opt/miniconda3/conda-bld/anaconda-anon-usage_1764636648062/work
anaconda-auth @ file:///opt/miniconda3/conda-bld/anaconda-cloud-auth-split_1765839823423/work
anaconda-cli-base @ file:///opt/miniconda3/conda-bld/anaconda-cli-base_1764888589923/work
annotated-doc==0.0.4
annotated-types @ file:///opt/miniconda3/conda-bld/annotated-types_1761744949290/work
anyio==4.13.0
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
argcomplete==3.6.3
boltons @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_97h34kncb9/croot/boltons_1751383736162/work
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1764961161081/work
certifi @ file:///opt/miniconda3/conda-bld/certifi_1763135329285/work/certifi
cffi @ file:///opt/miniconda3/conda-bld/cffi_1761832773185/work
charset-normalizer @ file:///opt/miniconda3/conda-bld/charset-normalizer_1761744840553/work
click @ file:///opt/miniconda3/conda-bld/click_1764332239180/work
conda @ file:///opt/miniconda3/conda-bld/conda_1765567809749/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ebkcs5yxh6/croot/conda-anaconda-tos_1755123330773/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1728603284288/work
conda-libmamba-solver @ file:///C:/miniconda3/conda-bld/conda-libmamba-solver_1764245615235/work/src
conda-package-handling @ file:///opt/miniconda3/conda-bld/conda-package-handling_1762366402479/work
conda_package_streaming @ file:///opt/miniconda3/conda-bld/conda-package-streaming_1762361629849/work
cryptography @ file:///opt/miniconda3/conda-bld/cryptography-split_1761931964095/work
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1728591855492/work
fastapi==0.135.3
frozendict @ file:///opt/miniconda3/conda-bld/frozendict_1761750665916/work
google-photos-takeout-helper==2.3.0
h11==0.16.0
idna @ file:///opt/miniconda3/conda-bld/idna_1761912005426/work
jaraco.classes @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_97dr4u6nc1/croot/jaraco.classes_1755516335437/work
jaraco.context @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jaraco.context_1731716684676/work
jaraco.functools @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_91zaq5_008/croot/jaraco.functools_1740408121405/work
jsonpatch @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jsonpatch_1728594556110/work
jsonpointer @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_bdid71leqg/croot/jsonpointer_1753788445166/work
keyring @ file:///opt/miniconda3/conda-bld/keyring_1763637309871/work
libmambapy @ file:///opt/miniconda3/conda-bld/mamba-split_1763111381025/work/libmambapy
loguru==0.6.0
markdown-it-py @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6av49xs_3l/croot/markdown-it-py_1756299201304/work
mdurl @ file:///opt/miniconda3/conda-bld/mdurl_1758552175942/work
menuinst @ file:///opt/miniconda3/conda-bld/menuinst_1765382369523/work
more-itertools @ file:///opt/miniconda3/conda-bld/more-itertools_1761121541654/work
msgpack @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0d73gp34a_/croot/msgpack-python_1750958625811/work
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079023/work
piexif==1.1.3
pipx==1.11.0
pkce @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pkce_1728592680499/work
platformdirs @ file:///opt/miniconda3/conda-bld/platformdirs_1762356666077/work
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_bei6q1cvvt/croot/pluggy_1733169621717/work
psutil==7.2.2
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_61k1o_oje7/croot/pycosat_1736868429741/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496048647/work
pydantic @ file:///opt/miniconda3/conda-bld/pydantic_1764083527102/work
pydantic-settings @ file:///opt/miniconda3/conda-bld/pydantic-settings_1764165151428/work
pydantic_core @ file:///opt/miniconda3/conda-bld/pydantic-core_1764009784214/work
Pygments @ file:///opt/miniconda3/conda-bld/pygments_1762431584667/work
PyJWT @ file:///opt/miniconda3/conda-bld/pyjwt_1764332099092/work
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753010927/work
python-dotenv @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0bqseadoho/croot/python-dotenv_1745610246418/work
readchar @ file:///opt/miniconda3/conda-bld/readchar_1760613300945/work
requests @ file:///opt/miniconda3/conda-bld/requests_1762359592010/work
rich @ file:///opt/miniconda3/conda-bld/rich_1760375480293/work
ruamel.yaml @ file:///opt/miniconda3/conda-bld/ruamel.yaml_1762535951282/work
ruamel.yaml.clib @ file:///opt/miniconda3/conda-bld/ruamel.yaml.clib_1762530070208/work
semver @ file:///opt/miniconda3/conda-bld/semver_1761903026306/work
setuptools==80.9.0
shellingham @ file:///opt/miniconda3/conda-bld/shellingham_1761912204150/work
speedtest-cli==2.1.3
SQLAlchemy==2.0.49
starlette==1.0.0
tomli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03b1up8uh3/croot/tomli_1753774598375/work
tqdm==4.64.0
truststore @ file:///opt/miniconda3/conda-bld/truststore_1762520923820/work
typer==0.20.0
typer-slim==0.20.0
typing-inspection @ file:///opt/miniconda3/conda-bld/typing-inspection_1760614169744/work
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
urllib3 @ file:///opt/miniconda3/conda-bld/urllib3_1765399201892/work
userpath==1.9.2
uvicorn==0.44.0
wheel==0.45.1
win32-setctime==1.1.0
zstandard @ file:///opt/miniconda3/conda-bld/zstandard_1758189078532/work
